example: extract widgets
